Correct `sample_size` of Stable Diffusion 2's unet to have correct width and height default
Since `diffusers==0.9.0` the width and height is automatically inferred from the `sample_size` attribute of your unet's config. It seems like your diffusion model has the same architecture as Stable Diffusion 2 which means that when using this model, by default an image size of 768x768 should be generated. This in turn means the unet's sample size should be **96**. 

 In order to suppress to update your configuration on the fly and to suppress the deprecation warning added in this PR: https://github.com/huggingface/diffusers/pull/1406/files#r1035703505 it is strongly recommended to merge this PR.